♻️ refactor: domain과 global 분리하기
