--- purithm/src/main/java/com/example/purithm/auth/dto/response/LoginResponseDto.java
@@ -1,12 +1,0 @@
-package com.example.purithm.auth.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class LoginResponseDto {
-  int code;
-  String message;
-  String token;
-}
