--- purithm/src/main/java/com/example/purithm/auth/dto/response/LoginSuccessResponseDto.java
@@ -1,10 +1,0 @@
-package com.example.purithm.auth.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public class LoginSuccessResponseDto {
-  int code;
-  String message;
-  String token;
-}
