--- purithm/src/main/java/com/example/purithm/auth/dto/response/SocialUserInfoDto.java
@@ -1,12 +1,0 @@
-package com.example.purithm.auth.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class SocialUserInfoDto {
-  private String nickname;
-  private String username;
-  private String profile;
-}
