--- purithm/src/main/java/com/example/purithm/auth/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,21 +1,0 @@
-package com.example.purithm.auth.handler;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-@Component
-public class OAuth2AuthenticationFailureHandler implements AuthenticationFailureHandler {
-
-  @Override
-  public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
-      AuthenticationException exception) throws IOException, ServletException {
-    response.setContentType("application/json");
-    response.setStatus(response.SC_UNAUTHORIZED);
-    response.getOutputStream().println("{\"error\": \"Unauthorized\", \"message\": \"" + exception.getMessage() + "\"}");
-  }
-}
