--- purithm/src/main/java/com/example/purithm/check/CheckController.java
@@ -1,15 +1,0 @@
-package com.example.purithm.check;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class CheckController {
-
-  @GetMapping("/api")
-  public void checkToken() {
-    return;
-  }
-
-
-}
