+++ purithm/src/main/java/com/example/purithm/domain/user/entity/User.java
@@ -1,0 +1,32 @@
+package com.example.purithm.domain.user.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  private String username;
+
+  private String nickname;
+
+  private String profile;
+
+  @Builder
+  public User(String username, String nickname, String profile) {
+    this.username = username;
+    this.nickname = nickname;
+    this.profile = profile;
+  }
+}
