+++ purithm/src/main/java/com/example/purithm/domain/user/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.example.purithm.domain.user.repository;
+
+import com.example.purithm.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+  User findByUsername(String username);
+}
