+++ purithm/src/main/java/com/example/purithm/global/auth/dto/response/LoginResponseDto.java
@@ -1,0 +1,12 @@
+package com.example.purithm.global.auth.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class LoginResponseDto {
+  int code;
+  String message;
+  String token;
+}
