+++ purithm/src/main/java/com/example/purithm/global/auth/dto/response/LoginSuccessResponseDto.java
@@ -1,0 +1,10 @@
+package com.example.purithm.global.auth.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public class LoginSuccessResponseDto {
+  int code;
+  String message;
+  String token;
+}
