+++ purithm/src/main/java/com/example/purithm/global/auth/entity/CustomOAuth2User.java
@@ -1,0 +1,35 @@
+package com.example.purithm.global.auth.entity;
+
+import java.util.Collection;
+import java.util.Map;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+public class CustomOAuth2User implements OAuth2User {
+
+  private final String username;
+
+  public CustomOAuth2User(String username) {
+    this.username = username;
+  }
+
+  @Override
+  public <A> A getAttribute(String name) {
+    return OAuth2User.super.getAttribute(name);
+  }
+
+  @Override
+  public Map<String, Object> getAttributes() {
+    return null;
+  }
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    return null;
+  }
+
+  @Override
+  public String getName() {
+    return username;
+  }
+}
