+++ purithm/src/main/java/com/example/purithm/global/auth/exception/JWTAuthenticationEntryPoint.java
@@ -1,0 +1,22 @@
+package com.example.purithm.global.auth.exception;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JWTAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+  @Override
+  public void commence(HttpServletRequest request, HttpServletResponse response,
+      AuthenticationException authException) throws IOException, ServletException {
+    response.sendError(
+        HttpServletResponse.SC_UNAUTHORIZED,
+        authException.getLocalizedMessage()
+    );
+  }
+}
