+++ purithm/src/main/java/com/example/purithm/global/auth/jwt/JWTUtil.java
@@ -1,0 +1,36 @@
+package com.example.purithm.global.auth.jwt;
+
+import io.jsonwebtoken.Jwts;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+import javax.crypto.SecretKey;
+import javax.crypto.spec.SecretKeySpec;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JWTUtil {
+  private SecretKey secretKey;
+
+  public JWTUtil(@Value("${spring.jwt.secret}")String secret) {
+    this.secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), Jwts.SIG.HS256.key().build().getAlgorithm());
+  }
+
+  public String getUsername(String token) {
+
+    return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("username", String.class);
+  }
+
+  public Boolean isExpired(String token) {
+    return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().getExpiration().before(new Date());
+  }
+
+  public String createJwt(String nickname, Long expiredMs) {
+    return Jwts.builder()
+        .claim("nickname", nickname)
+        .issuedAt(new Date(System.currentTimeMillis()))
+        .expiration(new Date(System.currentTimeMillis() + expiredMs))
+        .signWith(secretKey)
+        .compact();
+  }
+}
