+++ purithm/src/main/java/com/example/purithm/global/config/SwaggerConfig.java
@@ -1,0 +1,18 @@
+package com.example.purithm.global.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+  @Bean
+  public OpenAPI openAPI() {
+    return new OpenAPI()
+        .info(new Info()
+            .title("Purithm API")
+            .version("1.0.0")
+            .description("Purithm API 문서입니다."));
+  }
+}
