+++ purithm/src/main/java/com/example/purithm/global/config/WebClientConfig.java
@@ -1,0 +1,28 @@
+package com.example.purithm.global.config;
+
+import io.netty.channel.ChannelOption;
+import io.netty.handler.timeout.ReadTimeoutHandler;
+import java.net.http.HttpClient;
+import java.util.function.Function;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.client.reactive.ClientHttpConnector;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfig {
+
+  @Value("${spring.webflux.url}")
+  private String baseUrl;
+
+  @Bean
+  public WebClient webClient() {
+    return WebClient.builder()
+        .baseUrl(baseUrl)
+        .build();
+  }
+
+}
