--- purithm/src/main/java/com/example/purithm/user/entity/User.java
@@ -1,32 +1,0 @@
-package com.example.purithm.user.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  private String username;
-
-  private String nickname;
-
-  private String profile;
-
-  @Builder
-  public User(String username, String nickname, String profile) {
-    this.username = username;
-    this.nickname = nickname;
-    this.profile = profile;
-  }
-}
