--- purithm/src/main/java/com/example/purithm/user/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.example.purithm.user.repository;
-
-import com.example.purithm.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-  User findByUsername(String username);
-}
